+++ interchain/modules-clone-testing/src/common.rs
@@ -1,9 +1,9 @@
 use abstract_client::{AbstractClient, Environment};
 use abstract_interface::VCQueryFns;
 use abstract_interface::{Abstract, VCExecFns};
-use cosmwasm_std::{coins, Addr};
+use cosmwasm_std::coins;
+use cw_orch::daemon::ChainInfo;
 use cw_orch::prelude::*;
-use cw_orch::{contract::Deploy, daemon::ChainInfo};
 use cw_orch_clone_testing::CloneTesting;
 
 /// Returns a shared tokio runtime for all tests