+++ modules/contracts/adapters/dex/tests/common/mod.rs
@@ -1,6 +1,5 @@
 use abstract_core::objects::gov_type::GovernanceDetails;
 use abstract_interface::{AbstractAccount, AccountFactory};
-use cosmwasm_std::Addr;
 use cw_orch::prelude::*;
 pub fn create_default_account<Chain: CwEnv>(
     factory: &AccountFactory<Chain>,