+++ modules/contracts/adapters/dex/tests/raw.rs
@@ -14,7 +14,7 @@ use abstract_dex_adapter::interface::DexAdapter;
 use abstract_dex_standard::raw_action::DexRawAction;
 use abstract_interface::{Abstract, AbstractAccount};
 use common::create_default_account;
-use cosmwasm_std::{coin, Decimal, Empty};
+use cosmwasm_std::{coin, Decimal};
 use cw_orch::prelude::*;
 use speculoos::*;
 use wyndex_bundle::{EUR, USD, WYNDEX as WYNDEX_WITHOUT_CHAIN, WYNDEX_OWNER};