+++ modules/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Etf fee distribution fixed
+- Replaced empty enum migrate messages with empty structs
 
 ## [0.21.0] - 2024-02-20